requirements.txt
fastapi
uvicorn[standard]
beautifulsoup4
pandas
lxml
pydantic
psycopg2-binary
google-cloud-storage
cloud-sql-python-connector[pg8000]
langchain-community
google-cloud-aiplatform
langchain
openai
pypdf
chromadb
tiktoken
langchain-google-genai